New communication logic
Instead of waiting for the end of file, we are buffering output and
parse it as soon as we hit an action(s). This is more appropriate for
larger files and is also more interactive (which is our intention - to
communicate with agda interactively). All communications with agda are
done on our own thread (dedicated just for communication) - Don’t
forget to call main thread when working with UI stuff. Also, no need
for closing pipes now, just write a message and notification will
handle the rest. Last thing, notifications are called recursively e.g.
once we get notification we need to send
waitForDataInBackgroundAndNotify again.

Next step is to migrate all Terminal calls to this communication
channel.